[JE2-297] Change tab styles in ArrivalDetails.tsx
