+++ frontend/src/features/orders/containers/OrderDetails.tsx
@@ -21,57 +21,12 @@ import { capitalize } from '@/utils/capitalizeFirstLetter.ts'
 import ServicesTable from '@/components/Tables/ServicesTable.tsx'
 import CancelButton from '@/components/Buttons/CancelButton.tsx'
 import LogsAccordionView from '@/components/LogsAccordionView/LogsAccordionView.tsx'
-import { useCallback, useEffect, useState } from 'react'
+
 
 const OrderDetails = () => {
-  const { order, loading, open, openArchiveModal, handleArchive, setOpen, setOpenArchiveModal, tabs, setTabs, confirmCancelModalOpen, handleCancel, setConfirmCancelModalOpen  } =
+  const { order, loading, open, openArchiveModal, handleArchive, setOpen, setOpenArchiveModal, tabs, setTabs, confirmCancelModalOpen, handleCancel, setConfirmCancelModalOpen, paddingTop, heightTab } =
     useOrderDetails()
 
-  interface NavigatorUAData {
-    getHighEntropyValues(hints: string[]): Promise<{ [key: string]: string }>
-    platform: string
-    brands: { brand: string; version: string }[]
-    mobile: boolean
-  }
-
-  interface ExtendedNavigator extends Navigator {
-    userAgentData?: NavigatorUAData
-  }
-
-  const getOS = useCallback(async (navigator: ExtendedNavigator): Promise<string> => {
-    if (navigator.userAgentData) {
-      const uaData = await navigator.userAgentData.getHighEntropyValues(['platform'])
-      return uaData.platform
-    } else {
-      const userAgent = navigator.userAgent
-
-      if (/Windows NT/.test(userAgent)) {
-        return 'Windows'
-      } else if (/Mac OS X/.test(userAgent)) {
-        return 'Mac OS'
-      } else if (/Android/.test(userAgent)) {
-        return 'Android'
-      } else if (/iPhone|iPad|iPod/.test(userAgent)) {
-        return 'iOS'
-      } else if (/Linux/.test(userAgent)) {
-        return 'Linux'
-      } else {
-        return 'Unknown'
-      }
-    }
-  }, [])
-
-  const [os, setOS] = useState<string>('Detecting...')
-
-  useEffect(() => {
-    getOS(navigator as ExtendedNavigator).then(setOS)
-  }, [getOS])
-
-  console.log(os)
-
-
-  const paddingTop = os === 'Mac OS' ? 'pt-0' : os === 'Windows' ? 'pt-0': os === 'Android' ? 'pt-0' : 'pt-2'
-  const heightTab = os === 'Mac OS' ? 'h-[45px]' : os === 'Windows' ? 'h-[50px]' : os === 'Android' ? 'h-auto' : 'h-[45px]'
 
   return (
     <>
